SCRD-3737 Add some usability options for running the a-c-p
This patch adds some options when you are testing models. You can now
specify the branch to use to pull the various input models. Also, if you
have cloned them locally, it allows you to use your current branch
instead of reseting to master. The README.md was also updated to add these
changes and fix some obsoleted text. In keeping with standards, both
README files were added to the .copyrightignore

+++ Scripts/setup-ardana-cp.sh
@@ -24,32 +24,47 @@ clone_base="$(dirname "${acp_base}")"
 
 help ()
 {
-  echo "-g | --<git_base> Git base to use (default: $git_base)"
-  echo "-b | --branch <branch>  Git branch to use (default: $branch)"
-  echo "-n | --no-ref           Don't create reference"
-  echo "-p | --pull             Clone repos even if the already exist"
+  echo "-g | --<git_base>         Git base to use (default: $git_base)"
+  echo "-b | --branch <branch>    Git branch to use for a-c-p-repo (default: $branch)"
+  echo "-m | --m-branch <branch>  Git branch to use for a-i-m repo (default: $m_branch)"
+  echo "-r | --r-branch <branch>  Git branch to use for a-i-m-r repo (default: $r_branch)"
+  echo "-d | --d-branch <branch>  Git branch to use for ardana dcn repo (default: $d_branch)"
+  echo "-o | --o-branch <branch>  Git branch to use for ardana odl repo (default: $o_branch)"
+  echo "-n | --no-ref             Don't create reference"
+  echo "-p | --pull               Clone repos even if the already exist"
+  echo "-x | --no-reset           Do not reset head for locally cloned repos"
 }
 
-OPTS=$(getopt -o g:b:hnp --long git_base,branch,help,no-ref,pull -n 'parse-options' -- "$@")
+OPTS=$(getopt -o g:b:m:r:d:o:hnpx --long git_base,branch,m-branch,r-branch,d-branch,o-branch,help,no-ref,pull,no-reset -n 'parse-options' -- "$@")
 eval set -- "$OPTS"
 
 if [ $? != 0 ] ; then echo "Failed parsing options." >&2 ; exit 1 ; fi
 
 show_help=false
 branch="master"
+r_branch="master"
+m_branch="master"
+d_branch="master"
+o_branch="master"
 git_base="git://git.suse.provo.cloud"
 git_namespace="ardana"
 cp_opts=
 build_ref=true
 fresh_clone=false
+reset_head=false
 
 while true; do
   case $1 in
     -g | --git-base) shift; git_base=$1; shift;;
     -b | --branch) shift; branch=$1; shift ;;
+    -m | --m-branch) shift; m_branch=$1; shift ;;
+    -r | --r-branch) shift; r_branch=$1; shift ;;
+    -d | --d-branch) shift; d_branch=$1; shift ;;
+    -o | --o-branch) shift; o_branch=$1; shift ;;
     -h | --help) show_help=true; shift ;;
     -n | --no-ref) build_ref=false; shift ;;
     -p | --pull) fresh_clone=true; shift ;;
+    -x | --no-reset) reset_head=false; shift ;;
     -- ) shift; break ;;
     *) break;;
   esac
@@ -90,6 +105,10 @@ fi
 
 echo using branch $branch
 echo using external branch $external_branch
+echo using ardana-input-model branch $m_branch
+echo using ardana-input-model-ref branch $r_branch
+echo using ardana-extension-odl branch $o_branch
+echo using ardana-extension-dcn branch $d_branch
 
 target=$1
 mkdir $target
@@ -104,14 +123,17 @@ clone_repo_into_dir()
 
     if [[ -d "${clone_base}/${repo_name}" && $fresh_clone = "false" ]]
     then
-        # copy the repo and clear out any local
-        # uncommitted changes
+        # copy the repo and, if reset_head is true, clear
+        # out any local uncommitted changes
         cp -a "${clone_base}/${repo_name}" .
-        (
-            cd ${repo_name}
-            git reset --hard
-            git clean -x -d -f -f
-        )
+        if [[ $reset_head = "true" ]]
+        then
+            (
+                cd ${repo_name}
+                git reset --hard
+                git clean -x -d -f -f
+            )
+        fi
     else
         git clone -b "${branch}" "${git_base}/${repo}" "${dest_dir}"
     fi
@@ -122,18 +144,18 @@ cd $work_dir
 #
 # Clone the a-i-m repo
 #
-clone_repo_into_dir ${git_namespace}/ardana-input-model ${branch} ardana-input-model
+clone_repo_into_dir ${git_namespace}/ardana-input-model ${m_branch} ardana-input-model
 
 #
 # Clone the a-i-m-ref repo
 #
-clone_repo_into_dir ${git_namespace}/ardana-input-model-ref ${branch} ardana-input-model-ref
+clone_repo_into_dir ${git_namespace}/ardana-input-model-ref ${r_branch} ardana-input-model-ref
 
 #
 # Clone the extensions repos
 #
-clone_repo_into_dir ardana/ardana-extensions-dcn ${branch} ardana-extensions-dcn
-clone_repo_into_dir ardana/ardana-extensions-odl ${branch} ardana-extensions-odl
+clone_repo_into_dir ardana/ardana-extensions-dcn ${d_branch} ardana-extensions-dcn
+clone_repo_into_dir ardana/ardana-extensions-odl ${o_branch} ardana-extensions-odl
 
 # Look for any additional service definitons and copy them into
 # the input model directory so they get picked up.   Make sure that
